remove unused logic
