+++ Sources/CovidCertificateSDK/NationalRules/AcceptedProducts/AcceptedProducts.swift
@@ -14,7 +14,6 @@ import Foundation
 class AcceptedProducts {
     static var shared = AcceptedProducts()
 
-    private let acceptedTests: AcceptedTests
     private let acceptedVaccines: AcceptedVaccines
 
     // MARK: - Vaccination API
@@ -28,22 +27,9 @@ class AcceptedProducts {
         return entry?.totalDosisNumber
     }
 
-    // MARK: - Test API
-
-    func testIsAccepted(test: Test) -> Bool {
-        if test.type == TestType.Pcr.rawValue {
-            return true
-        }
-        return acceptedTests.entries.contains {
-            return test.ratTestNameAndManufacturer == $0.key
-        }
-    }
-
     // MARK: - Init
 
     init() {
-        let acceptedTestsJson = try! String(contentsOf: Bundle.module.url(forResource: "test-manf", withExtension: "json")!)
-        acceptedTests = try! JSONDecoder().decode(AcceptedTests.self, from: acceptedTestsJson.data(using: .utf8)!)
         let acceptedVaccinesJson = try! String(contentsOf: Bundle.module.url(forResource: "accepted-vaccines", withExtension: "json")!)
         acceptedVaccines = try! JSONDecoder().decode(AcceptedVaccines.self, from: acceptedVaccinesJson.data(using: .utf8)!)
     }
